LayoutTests/fast/css-grid-layout/grid-columns-rows-get-set-multiple-expected.txt

Test that setting and getting grid-definition-columns and grid-definition-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting |grid-definition-columns| and |grid-definition-rows| set through CSS
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-definition-columns') is "7px 11px"
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-definition-rows') is "17px 2px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-definition-columns') is "424px 792px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-definition-rows') is "162px 312px"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-definition-columns') is "0px 17px"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-definition-rows') is "0px 3px"
PASS window.getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-definition-columns') is "100px 120px"
PASS window.getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-definition-rows') is "150px 170px"
PASS window.getComputedStyle(gridWithThreeItems, '').getPropertyValue('-webkit-grid-definition-columns') is "15px 0px 100px"
PASS window.getComputedStyle(gridWithThreeItems, '').getPropertyValue('-webkit-grid-definition-rows') is "120px 18px 0px"
PASS window.getComputedStyle(gridWithPercentAndViewportPercent, '').getPropertyValue('-webkit-grid-definition-columns') is "400px 120px"
PASS window.getComputedStyle(gridWithPercentAndViewportPercent, '').getPropertyValue('-webkit-grid-definition-rows') is "210px 168px"
PASS window.getComputedStyle(gridWithFitContentAndFitAvailable, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithFitContentAndFitAvailable, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithMinMaxContent, '').getPropertyValue('-webkit-grid-definition-columns') is "0px 0px"
PASS window.getComputedStyle(gridWithMinMaxContent, '').getPropertyValue('-webkit-grid-definition-rows') is "0px 0px"
PASS window.getComputedStyle(gridWithMinMaxContentWithChildrenElement, '').getPropertyValue('-webkit-grid-definition-columns') is "7px 17px"
PASS window.getComputedStyle(gridWithMinMaxContentWithChildrenElement, '').getPropertyValue('-webkit-grid-definition-rows') is "11px 3px"
PASS window.getComputedStyle(gridWithMinMaxAndFixed, '').getPropertyValue('-webkit-grid-definition-columns') is "240px 15px"
PASS window.getComputedStyle(gridWithMinMaxAndFixed, '').getPropertyValue('-webkit-grid-definition-rows') is "120px 210px"
PASS window.getComputedStyle(gridWithMinMaxAndMinMaxContent, '').getPropertyValue('-webkit-grid-definition-columns') is "240px 15px"
PASS window.getComputedStyle(gridWithMinMaxAndMinMaxContent, '').getPropertyValue('-webkit-grid-definition-rows') is "120px 210px"
PASS window.getComputedStyle(gridWithFractionFraction, '').getPropertyValue('-webkit-grid-definition-columns') is "320px 480px"
PASS window.getComputedStyle(gridWithFractionFraction, '').getPropertyValue('-webkit-grid-definition-rows') is "225px 375px"
PASS window.getComputedStyle(gridWithFractionMinMax, '').getPropertyValue('-webkit-grid-definition-columns') is "45px 755px"
PASS window.getComputedStyle(gridWithFractionMinMax, '').getPropertyValue('-webkit-grid-definition-rows') is "586px 14px"

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is 'none'

Test getting and setting grid-definition-rows and grid-definition-columns through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "18px 22px"
PASS element.style.webkitGridDefinitionColumns is "18px 22px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "66px 70px"
PASS element.style.webkitGridDefinitionRows is "66px 70px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "440px 640px"
PASS element.style.webkitGridDefinitionColumns is "55% 80%"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "240px 378px"
PASS element.style.webkitGridDefinitionRows is "40% 63%"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "0px 0px"
PASS element.style.webkitGridDefinitionColumns is "auto auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "0px 0px"
PASS element.style.webkitGridDefinitionRows is "auto auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "0px 160px 22px"
PASS element.style.webkitGridDefinitionColumns is "auto 16em 22px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "336px 100px 0px"
PASS element.style.webkitGridDefinitionRows is "56% 10em auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "160px 20px"
PASS element.style.webkitGridDefinitionColumns is "16em minmax(16px, 20px)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "90px 0px"
PASS element.style.webkitGridDefinitionRows is "minmax(10%, 15%) auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "160px 640px"
PASS element.style.webkitGridDefinitionColumns is "16em 2fr"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "600px 0px"
PASS element.style.webkitGridDefinitionRows is "14fr auto"

Test getting wrong values set from CSS
PASS window.getComputedStyle(gridWithNoneAndAuto, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithNoneAndAuto, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithNoneAndFixed, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithNoneAndFixed, '').getPropertyValue('-webkit-grid-definition-rows') is "none"

Test setting and getting wrong values from JS
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"

Test setting grid-definition-columns and grid-definition-rows to 'inherit' through JS
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "50px 750px last"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "20px middle 45px"

Test setting grid-definition-columns and grid-definition-rows to 'initial' through JS
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "1200px middle 55px"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "200px line 400px line"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
